+++ PersistenciaRestaurante/src/main/java/interfaces/IClienteFrecuenteDAO.java
@@ -16,5 +16,5 @@ public interface IClienteFrecuenteDAO {
 
     public ClienteFrecuente registrarNuevoClienteFrecuente (ClienteFrecuente clienteFrecuente) throws PersistenciaException;
     public List<ClienteFrecuente> obtenerClientesFrecuentes(String filtro, String dato) throws PersistenciaException;
-    public ClienteFrecuente obtenerClientePorId(int id) throws PersistenciaException;
+    public ClienteFrecuente obtenerClientePorId(Long id) throws PersistenciaException;
 }
